[리팩토링] slice - actions, reducer 통합
