+++ src/components/LoginFormContainer.jsx
@@ -7,7 +7,7 @@ import {
   changeLoginField,
   requestLogin,
   logout,
-} from '../actions';
+} from '../slice';
 
 import { get } from '../utils';
